Adding customized packages GpGp2 and RobustGaSP2
GpGp2 has the integrated scaled Vecchia likelihood implemented along with the gradient for the integrated likelihood. Additionally, scaled Vecchia has been changed to a productwise covariance function across the input dimensions to match up with how RobustGaSP computes the covariance matrix. These are the main changes from the original GpGp package. 

RobustGaSP2 has been changed to accept Vecchia approximations (only for 1 dimensional outputs as of this commit) for the function ppgasp. Currently this only works with the default prior and the Matérn 3/2 covariance function.